attendance-service/requirements.txt
pyjwt
flask
psycopg2-binary==2.9.9
grpcio
requests
flask_cors